requirements.txt
#Core ZenML
zenml[server]>=0.53.0

#Paper ingestion
arxiv==1.4.7

#NLP & Embeddings
sentence-transformers==2.2.2
torch>=1.13.1
transformers>=4.31.0

#vector DB
qdrant-client==1.6.2

#Data Processing
numpy>=1.24.0
pandas>=2.0.0

#Optional 
sqlalchemy>=1.4.0

#Misc
tqdm

#UI
streamlit
#env
python-dotenv
